improved css and also added ai review

+++ src/backend/controllers/problem.controller.js
@@ -9,7 +9,12 @@ import generateFile from '../runSubmission/generateFile.js';
 import  execute from '../runSubmission/execute.js';
 import fs from 'fs';
 import { runCode } from '../runSubmission/execute.js';
+import path from 'path';   
+import { fileURLToPath } from 'url';
 
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const outputPath = path.join(__dirname, 'outputs');     
 
 const getAllProblems = asyncHandler(async (req, res) => {
        console.log('Fetching all problems');
@@ -17,7 +22,7 @@ const getAllProblems = asyncHandler(async (req, res) => {
 
              console.log('Fetching all problems');
           // Get all problems from the database
-             const problems = await Problem.find({}).select("problemId title").lean();
+             const problems = await Problem.find({}).select("problemId title difficultyLevel tags").lean();
             console.log(problems);
     
         //   Check if problems exist
@@ -144,6 +149,49 @@ const updateProblem=asyncHandler(async (req , res) => {
     }
 });
 
+
+const addTestCase = asyncHandler(async (req, res) => {
+    console.log('Adding test case');
+    try {
+        const { problemId } = req.params;
+        console.log(problemId);
+        const { input, output } = req.body;
+        console.log(input, output);
+        const problem = await Problem.findOne({ problemId });
+        if (!problem) {
+            console.log('Problem not found');
+            return res.status(404).json(new ApiResponse(404, 'Problem not found'));
+        }
+        const _id = problem._id;
+        console.log(_id);
+        
+        if (!input || !output) {
+            return res.status(400).json(new ApiResponse(400, 'Input and output are required'));
+        }
+      
+        const testCaseData = {
+            "problemId": _id,
+            "input": input,
+            "output": output,
+        };
+
+        const newTestCase = new TestCase(testCaseData);
+        await newTestCase.save();
+        if (!problem.testCases) {
+            problem.testCases = [];
+        }
+        
+        problem.testCases.push(newTestCase._id);
+        await problem.save();
+        console.log('Test case added successfully',problem.testCases.length);
+        return res.status(201).json(new ApiResponse(201, 'Test case added successfully', newTestCase));
+    } catch (error) {
+         console.error('Error adding test case:', error);
+        return res.status(500).json(new ApiResponse(500, 'Server error', { error: error.message }));
+    }
+});
+
+
 const submitProblem=asyncHandler(async (req , res) => {
     console.log('Submitting problem');
    
@@ -151,7 +199,7 @@ const submitProblem=asyncHandler(async (req , res) => {
        
         const { problemId } = req.params;
         const { code, lang } = req.body;
-    
+        console.log(problemId, code, lang);
         if (!code || !lang) {
             return res.status(400).json(new ApiResponse(400, 'Code and language are required'));
         }
@@ -168,9 +216,11 @@ const submitProblem=asyncHandler(async (req , res) => {
         const testCasesId = problem?.testCases || [];
        
         let passedCount = 0;
-         console.log('Test cases:', testCasesId.length);
+        
          const filePath = await generateFile(lang, code);
          const testCases = await TestCase.find({ _id: { $in: testCasesId } }).select('input output').lean();
+         let error=null;
+   
 for (let index = 0; index < testCasesId.length; index++) {
    const testCase=testCases[index];
     console.log(testCase.input);
@@ -179,17 +229,20 @@ for (let index = 0; index < testCasesId.length; index++) {
     const output = await runCode({ filePath, inputPath });
     console.log('Output:', output);
   
-    fs.unlinkSync(inputPath, (err) => {
-        if (err) {
-            console.error('Error deleting file:', err);
-        } else {
-            console.log('File deleted successfully');
-        }
-    });
-    // Assuming testCases[index].output is an array of accepted outputs
-    if (testCase.output.includes(output.trim())) {
+  
+    if (testCase.output.includes(output)) {
         passedCount++;
     }
+    if(output?.error)
+    {
+        error=output.error;
+        
+        break;
+    }
+}
+if(error=='Time Limit Exceeded')
+{
+    return res.status(400).json(new ApiResponse(400, 'Time Limit Exceeded', { error }));
 }
 fs.unlinkSync(filePath, (err) => {
     if (err) {
@@ -198,11 +251,15 @@ fs.unlinkSync(filePath, (err) => {
         console.log('File deleted successfully');
     }
 });
+
 submissionData.output = `${passedCount}/${testCasesId.length} test cases passed`;
 const submission = new Submission(submissionData);
 await submission.save();
 
 if (passedCount === testCasesId.length) {
+    const user= req.user;
+    user.problemSolved.push(problemId);
+    await user.save();
     return res.status(200).json(new ApiResponse(200, 'Accepted', { submissionId: submission._id, output: submission.output }));
 } else {
     return res.status(400).json(new ApiResponse(400, 'Wrong Answer', { passed: passedCount }));
@@ -258,47 +315,6 @@ const runProblem=asyncHandler(async (req , res) => {
     catch
     (error) {
         console.error('Error running problem:', error);
-        return res.status(500).json(new ApiResponse(500, 'Server error', { error: error.message }));
-    }
-});
-
-const addTestCase = asyncHandler(async (req, res) => {
-    console.log('Adding test case');
-    try {
-        const { problemId } = req.params;
-        console.log(problemId);
-        const { input, output } = req.body;
-        console.log(input, output);
-        const problem = await Problem.findOne({ problemId });
-        if (!problem) {
-            console.log('Problem not found');
-            return res.status(404).json(new ApiResponse(404, 'Problem not found'));
-        }
-        const _id = problem._id;
-        console.log(_id);
-        
-        if (!input || !output) {
-            return res.status(400).json(new ApiResponse(400, 'Input and output are required'));
-        }
-      
-        const testCaseData = {
-            "problemId": _id,
-            "input": input,
-            "output": output,
-        };
-
-        const newTestCase = new TestCase(testCaseData);
-        await newTestCase.save();
-        if (!problem.testCases) {
-            problem.testCases = [];
-        }
-        
-        problem.testCases.push(newTestCase._id);
-        await Problem.updateOne({ testCases: problem.testCases });
-        console.log('Test case added successfully',problem.testCases.length);
-        return res.status(201).json(new ApiResponse(201, 'Test case added successfully', newTestCase));
-    } catch (error) {
-         console.error('Error adding test case:', error);
         return res.status(500).json(new ApiResponse(500, 'Server error', { error: error.message }));
     }
 });
